Fixed bug: The download started with a faster pace, and became very slow at the end.
Reason:
We split the images equally between the processes.
If one process received many lightweight images to download,
while another process received many heavyweight image,
the former process would finish earlier than the latter.
This also means the parallelism could be improved.

Solution:
Changed the split of the images in the following way:
Instead of prior splitting the images between the processes, we
will send them each 1 image at a time, and when they are done, they will receive another one.
This way the distribution is the best.
(Although we now get more delay due to the more function calls)